+++ daemon/src/engine/client/cl_cgame.cpp
@@ -44,7 +44,7 @@ Maryland 20850 USA.
 #include "framework/CommandSystem.h"
 #include "framework/CvarSystem.h"
 #include "framework/CrashDump.h"
-#include "framework/Profiler.h"
+#include "../common/Profiler.h"
 
 #define __(x) Trans_GettextGame(x)
 #define C__(x, y) Trans_PgettextGame(x, y)
