+++ daemon/src/engine/client/cl_scrn.cpp
@@ -36,7 +36,7 @@ Maryland 20850 USA.
 
 #include "client.h"
 #include "qcommon/q_unicode.h"
-#include "framework/Profiler.h"
+#include "../common/Profiler.h"
 
 bool scr_initialized; // ready to draw
